requirements.txt
pillow
numpy
requests
spotipy
RGBMatrixEmulator
mfrc522-python
